Code style: Decompose ClassReflection to class files according to PSR-4
