+++ src/TsLinkPhp/ClassReflection.php
@@ -5,8 +5,6 @@ declare(strict_types=1);
 namespace Murdej\TsLinkPhp;
 
 use ReflectionClass;
-use ReflectionNamedType;
-use ReflectionUnionType;
 
 class ClassReflection
 {
@@ -96,42 +94,4 @@ class ClassReflection
             $this->parseClass($className);
         }
     }
-}
-
-class ClassReflectionMethod
-{
-    public string $name;
-
-    /** @var ClassReflectionMethodParam[] */
-    public array $params;
-
-    public ReflectionNamedType|ReflectionUnionType|null|string $returnDataType;
-
-    public bool $rawResult = false;
-
-    public bool $newResult = false;
-
-    public function prepare()
-    {
-        if (is_string($this->returnDataType) && str_starts_with($this->returnDataType, 'new ')) {
-            $this->returnDataType = substr($this->returnDataType, 4);
-            $this->newResult = true;
-        }
-    }
-
-    public function getCallOpts(): array
-    {
-        return [
-            'rawResult' => $this->rawResult,
-        ];
-    }
-}
-
-class ClassReflectionMethodParam
-{
-    public string $name;
-
-    public bool $nullable;
-
-    public ReflectionNamedType|ReflectionUnionType|null|string $dataType;
 }
